caught a div/0
For variable load types the class performs a calculation to compute the centroid of the load area which contains a w1+w2 denominator term. When w1 = -1*w2 this results in a div/0 error. Applied loads of this type are now caught and split into the positive and negative component loads.