div/0 catch
caught a div/0 error when a distributed load was applied exactly from one interior support to another interior support. Logic was resulting in trying to split the load exactly at the left support resulting in the a and b values being equal and the slope formula for a variable load relies on (w2-w1)/c, where c = b-a.