Computation Time Improvement
Revised the method of computing the root results to occur during the flexibility solution process, for a 6 span beam which performs full load patterning this results in performing 11 computations for each load combination and pattern in lieu of the upwards of 7,000 additional charting stations being done previously. 
The root results are now also saved as a class variable dictionary for easy reference and reporting in the general results tab and no longer just on the charts.